Moved api gateway to local files
