+++ dev-portal/package.json
@@ -11,7 +11,9 @@
   "dependencies": {
     "amazon-cognito-identity-js": "^2.0.3",
     "aws-sdk": "^2.235.1",
+    "axios": "^0.18.0",
     "chart.js": "^2.4.0",
+    "crypto-js": "^3.1.9-1",
     "react": "^16.3.0",
     "react-datepicker": "^1.4.1",
     "react-dom": "^16.3.0",
@@ -19,7 +21,8 @@
     "react-router-dom": "^4.2.2",
     "semantic-ui-css": "^2.3.1",
     "semantic-ui-react": "^0.80.0",
-    "swagger-ui": "^3.14.1"
+    "swagger-ui": "^3.14.1",
+    "url-template": "^2.0.8"
   },
   "scripts": {
     "analyze": "source-map-explorer build/static/js/main.*",
