+++ dev-portal/public/index.html
@@ -5,19 +5,6 @@
     <!-- <meta name="viewport" content="width=device-width, initial-scale=1"> -->
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
 
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/lib/axios/dist/axios.standalone.js"></script>
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/lib/CryptoJS/rollups/hmac-sha256.js"></script>
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/lib/CryptoJS/rollups/sha256.js"></script>
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/lib/CryptoJS/components/hmac.js"></script>
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/lib/CryptoJS/components/enc-base64.js"></script>
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/lib/url-template/url-template.js"></script>
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/lib/apiGatewayCore/sigV4Client.js"></script>
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/lib/apiGatewayCore/apiGatewayClient.js"></script>
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/lib/apiGatewayCore/simpleHttpClient.js"></script>
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/lib/apiGatewayCore/utils.js"></script>
-    <script type="text/javascript" src="%PUBLIC_URL%/apigateway-js-sdk/apigClient.js"></script>
-    <script type="text/javascript" src="https://www.google.com/recaptcha/api.js" async defer></script>
-
     <!--
       Notice the use of %PUBLIC_URL% in the tag above.
       It will be replaced with the URL of the `public` folder during the build.
